third_party/blink/web_tests/external/wpt/css/css-animations/event-dispatch.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Active -> Idle, setting Animation.timeline = null
  assert_approx_equals: expected 0.1 +/- 0.0005 but got 0
Harness: the test ran to completion.
